+++ src/App.tsx
@@ -1,31 +1,9 @@
-import { useEffect, useState } from 'react';
 import { HeroCard } from './components/HeroCard';
 import { Header } from './container/Header';
-import type { Weather } from './types/apiResponseType';
+import { useGetWeather } from './hooks/useGetWeather';
 
 function App() {
-  const [weather, setWeather] = useState<Weather | null>(null);
-
-  const getWeather = async () => {
-    try {
-      const res = await fetch(
-        `https://goweather.xyz/v2/weather/surkhet?unit=u`,
-        {
-          method: 'GET',
-        }
-      );
-
-      const data: Weather = await res.json();
-      setWeather(data);
-    } catch (err) {
-      console.log(err?.response);
-    }
-  };
-
-  useEffect(() => {
-    getWeather();
-  }, []);
-
+  const { weather } = useGetWeather();
   console.log(weather, '---weather here');
 
   return (
@@ -33,12 +11,10 @@ function App() {
       <div className='w-full'>
         <Header />
       </div>
-      <HeroCard temperature={weather?.temperature} />
-      <button
-        className='text-white bg-blue-500 p-2 rounded-md'
-        onClick={getWeather}>
-        Get Weather
-      </button>
+      <HeroCard
+        temperature={weather?.temperature}
+        wind={weather?.wind}
+      />
     </div>
   );
 }
